Group_Assignment_2/requirements.txt
biopython==1.83
certifi==2024.2.2
charset-normalizer==3.3.2
contourpy==1.2.1
cycler==0.12.1
et-xmlfile==1.1.0
fonttools==4.51.0
idna==3.7
kiwisolver==1.4.5
matplotlib==3.8.4
networkx==3.3
numpy==1.26.4
openpyxl==3.1.2
packaging==24.0
pandas==2.2.1
pillow==10.3.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.31.0
scipy==1.13.0
six==1.16.0
tzdata==2024.1
uniprot==1.3
urllib3==2.2.1
clustalo==1.2.4-7
libargtable2-0 (>= 7)
libc6 (>= 2.29)
libgcc-s1 (>= 3.0)
libgomp1 (>= 6)
libstdc++6 (>= 5.2)
pydot==2.0.0
pyparsing==3.1.2

